+++ hphp/runtime/ext/reflection/ext_reflection.cpp
@@ -998,7 +998,7 @@ static bool HHVM_METHOD(ReflectionMethod, isConstructor) {
   return isConstructor(func);
 }
 
-static int HHVM_METHOD(ReflectionMethod, getModifiers) {
+static int64_t HHVM_METHOD(ReflectionMethod, getModifiers) {
   auto const func = ReflectionFuncHandle::GetFuncFor(this_);
   return get_modifiers(func->attrs(), false, false);
 }
@@ -1201,7 +1201,7 @@ static String HHVM_METHOD(ReflectionClass, getEnumUnderlyingType) {
   return StrNR(cls->preClass()->enumBaseTy().typeName());
 }
 
-static int HHVM_METHOD(ReflectionClass, getModifiers) {
+static int64_t HHVM_METHOD(ReflectionClass, getModifiers) {
   auto const cls = ReflectionClassHandle::GetClassFor(this_);
   return get_modifiers(cls->attrs(), true, false);
 }
@@ -1964,7 +1964,7 @@ static bool HHVM_METHOD(ReflectionProperty, isDefault) {
   }
 }
 
-static int HHVM_METHOD(ReflectionProperty, getModifiers) {
+static int64_t HHVM_METHOD(ReflectionProperty, getModifiers) {
   auto const data = Native::data<ReflectionPropHandle>(this_);
   switch (data->getType()) {
     case ReflectionPropHandle::Type::Instance:
