+++ hphp/runtime/ext/sockets/ext_sockets.h
@@ -24,25 +24,25 @@
 namespace HPHP {
 ///////////////////////////////////////////////////////////////////////////////
 
-bool socket_create_pair_impl(int domain, int type, int protocol, Variant& fd,
+bool socket_create_pair_impl(int64_t domain, int64_t type, int64_t protocol, Variant& fd,
                              bool asStream);
 
 Variant HHVM_FUNCTION(socket_create,
-                      int domain,
-                      int type,
-                      int protocol);
+                      int64_t domain,
+                      int64_t type,
+                      int64_t protocol);
 Variant HHVM_FUNCTION(socket_create_listen,
-                      int port,
-                      int backlog = 128);
+                      int64_t port,
+                      int64_t backlog = 128);
 bool HHVM_FUNCTION(socket_create_pair,
-                   int domain,
-                   int type,
-                   int protocol,
+                   int64_t domain,
+                   int64_t type,
+                   int64_t protocol,
                    Variant& fd);
 Variant HHVM_FUNCTION(socket_get_option,
                       const Resource& socket,
-                      int level,
-                      int optname);
+                      int64_t level,
+                      int64_t optname);
 bool HHVM_FUNCTION(socket_getpeername,
                    const Resource& socket,
                    Variant& address,
@@ -57,29 +57,29 @@ bool HHVM_FUNCTION(socket_set_nonblock,
                    const Resource& socket);
 bool HHVM_FUNCTION(socket_set_option,
                    const Resource& socket,
-                   int level,
-                   int optname,
+                   int64_t level,
+                   int64_t optname,
                    const Variant& optval);
 bool HHVM_FUNCTION(socket_connect,
                    const Resource& socket,
                    const String& address,
-                   int port = 0);
+                   int64_t port = 0);
 bool HHVM_FUNCTION(socket_bind,
                    const Resource& socket,
                    const String& address,
-                   int port = 0);
+                   int64_t port = 0);
 bool HHVM_FUNCTION(socket_listen,
                    const Resource& socket,
-                   int backlog = 0);
+                   int64_t backlog = 0);
 Variant HHVM_FUNCTION(socket_select,
                       Variant& read,
                       Variant& write,
                       Variant& except,
                       const Variant& vtv_sec,
-                      int tv_usec = 0);
+                      int64_t tv_usec = 0);
 Variant HHVM_FUNCTION(socket_server,
                       const String& hostname,
-                      int port,
+                      int64_t port,
                       Variant& errnum,
                       Variant& errstr);
 Variant HHVM_FUNCTION(socket_accept,
@@ -87,7 +87,7 @@ Variant HHVM_FUNCTION(socket_accept,
 Variant HHVM_FUNCTION(socket_read,
                       const Resource& socket,
                       int64_t length,
-                      int type = 0);
+                      int64_t type = 0);
 Variant HHVM_FUNCTION(socket_write,
                       const Resource& socket,
                       const String& buffer,
@@ -96,33 +96,33 @@ Variant HHVM_FUNCTION(socket_send,
                       const Resource& socket,
                       const String& buf,
                       int64_t len,
-                      int flags);
+                      int64_t flags);
 Variant HHVM_FUNCTION(socket_sendto,
                       const Resource& socket,
                       const String& buf,
                       int64_t  len,
-                      int flags,
+                      int64_t flags,
                       const String& addr,
-                      int port = -1);
+                      int64_t port = -1);
 Variant HHVM_FUNCTION(socket_recv,
                       const Resource& socket,
                       Variant& buf,
                       int64_t len,
-                      int flags);
+                      int64_t flags);
 Variant HHVM_FUNCTION(socket_recvfrom,
                       const Resource& socket,
                       Variant& buf,
                       int64_t len,
-                      int flags,
+                      int64_t flags,
                       Variant& name,
                       Variant& port);
 bool HHVM_FUNCTION(socket_shutdown,
                    const Resource& socket,
-                   int how = 0);
+                   int64_t how = 0);
 void HHVM_FUNCTION(socket_close,
                    const Resource& socket);
 String HHVM_FUNCTION(socket_strerror,
-                     int errnum);
+                     int64_t errnum);
 int64_t HHVM_FUNCTION(socket_last_error,
                       const Variant& socket = uninit_variant);
 void HHVM_FUNCTION(socket_clear_error,
@@ -130,10 +130,10 @@ void HHVM_FUNCTION(socket_clear_error,
 Variant HHVM_FUNCTION(getaddrinfo,
                       const String& host,
                       const String& port,
-                      int family = 0,
-                      int socktype = 0,
-                      int protocol = 0,
-                      int flags = 0);
+                      int64_t family = 0,
+                      int64_t socktype = 0,
+                      int64_t protocol = 0,
+                      int64_t flags = 0);
 
 ///////////////////////////////////////////////////////////////////////////////
 // Zend defines the following in network, so these are not part of the
@@ -141,13 +141,13 @@ Variant HHVM_FUNCTION(getaddrinfo,
 
 Variant HHVM_FUNCTION(fsockopen,
                       const String& hostname,
-                      int port,
+                      int64_t port,
                       Variant& errnum,
                       Variant& errstr,
                       double timeout = -1.0);
 Variant HHVM_FUNCTION(pfsockopen,
                       const String& hostname,
-                      int port,
+                      int64_t port,
                       Variant& errnum,
                       Variant& errstr,
                       double timeout = -1.0);
@@ -155,7 +155,7 @@ Variant HHVM_FUNCTION(pfsockopen,
 ///////////////////////////////////////////////////////////////////////////////
 
 Variant socket_server_impl(const HostURL &hosturl,
-                           int flags,
+                           int64_t flags,
                            Variant& errnum,
                            Variant& errstr,
                            const Variant& context = uninit_variant);
@@ -165,4 +165,3 @@ Variant sockopen_impl(const HostURL &hosturl, Variant& errnum,
 
 ///////////////////////////////////////////////////////////////////////////////
 }
-
